pyproject.toml
[tool.poetry]
name = "jordgubbe"
version = "0.1.0"
description = ""
authors = ["Joar Wandborg <joar@wandborg.se>"]

[tool.poetry.dependencies]
python = "^3.7"
strawberry-graphql = {path = "./strawberry/dist/strawberry-graphql-0.12.0.tar.gz"}
[tool.poetry.dev-dependencies]
mypy = "^0.711.0"
ipython = "^7.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
